Added a dedicated NamedList class to mimic R's named lists. (#10)
This allows integer indexing, slicing, etc. while also adding dict-like behavior
with getting/setting by (string) key. It also supports multiple instances of the
same name, which is consistent with R's behavior.

This new class is supported by an additional Names class, which is basically
a StringList but disallowing Nones and having more behind-the-scenes code
for efficiently mapping a name to a positional index.