Overhaul implementation of NamedLists and StringLists. (#12)
We no longer inherit from a list, so as to avoid complications with the base
types. We also refactor the StringList class so that it is now based on
NamedList, thus giving it access to names. (We make sure to prevent
naming of Factor levels, though, as that is unnecessarily complicated.)