Overhauled the Names class so that it no longer inherits from list. (#14)
We instead mimic the list interface so that users get the same experience,
without us having to actually abide by all expectations of the list data
structure. In particular, it means that we now can provide functional versions
of all methods to allow callers to apply modifications without side effects.