Added new Factor method to replace levels without remapping codes.
The new replace_levels() method is much simpler and more consistent with the
levels()<- method in R. The existing set_levels() functionality is moved to the
new remap_levels() to avoid confusion; set_levels() itself is now an alias to
either remap_levels() or replace_levels(), depending on the remap= argument.
The default remap=True has a deprecation warning as I hope to change it to
False, such that set_levels() is eventually the same as replace_levels().